requirements.txt
flasgger==0.9.7.1
Flask==3.1.0
Flask_Cors==5.0.0
msgpack_python==0.5.6
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.1
Werkzeug==3.1.3
